+++ fog/ledger/server/src/bin/router.rs
@@ -4,6 +4,7 @@ use std::env;
 
 use clap::Parser;
 use mc_common::logger::log;
+use mc_fog_block_provider::{BlockProvider, LocalBlockProvider, MobilecoindBlockProvider};
 use mc_fog_ledger_enclave::{LedgerSgxEnclave, ENCLAVE_FILE};
 use mc_fog_ledger_server::{LedgerRouterConfig, LedgerRouterServer};
 use mc_ledger_db::LedgerDB;
@@ -43,30 +44,49 @@ fn main() {
     let enclave = LedgerSgxEnclave::new(
         enclave_path,
         &config.client_responder_id,
-        config.omap_capacity,
+        // The router doesn't use the OMAP, so we can set it to 0.
+        0,
         logger.clone(),
     );
 
-    let ledger_db = LedgerDB::open(&config.ledger_db).expect("Could not read ledger DB");
-    let watcher_db =
-        WatcherDB::open_ro(&config.watcher_db, logger.clone()).expect("Could not open watcher DB");
+    let (block_provider, ledger_db) = match (
+        config.ledger_db.as_ref(),
+        config.watcher_db.as_ref(),
+        config.mobilecoind_uri.as_ref(),
+    ) {
+        (Some(ledger_db_path), Some(watcher_db_path), None) => {
+            let ledger_db = LedgerDB::open(ledger_db_path).expect("Could not read ledger DB");
+            let watcher = WatcherDB::open_ro(watcher_db_path, logger.clone())
+                .expect("Could not open watcher DB");
 
-    let mut router_server = LedgerRouterServer::new(
-        config,
-        enclave,
-        ledger_db.clone(),
-        watcher_db,
-        logger.clone(),
-    );
+            (
+                LocalBlockProvider::new(ledger_db.clone(), watcher) as Box<dyn BlockProvider>,
+                Some(ledger_db),
+            )
+        }
+
+        (None, None, Some(mobilecoind_uri)) => (
+            MobilecoindBlockProvider::new(mobilecoind_uri, &logger) as Box<dyn BlockProvider>,
+            None,
+        ),
+
+        _ => panic!("invalid configuration, need either ledger_db+watcher_db or mobilecoind_uri"),
+    };
+
+    let mut router_server =
+        LedgerRouterServer::new(config, enclave, block_provider, logger.clone());
     router_server.start();
 
     loop {
-        // The ledger database is read by this service, but updated by another service.
-        // In order to keep this service's metrics up to date, we need to update them
-        // periodically.
-        if let Err(e) = ledger_db.update_metrics() {
-            log::error!(logger, "Error updating ledger metrics: {:?}", e);
+        if let Some(ledger_db) = ledger_db.as_ref() {
+            // The ledger database is read by this service, but updated by another service.
+            // In order to keep this service's metrics up to date, we need to update them
+            // periodically.
+            if let Err(e) = ledger_db.update_metrics() {
+                log::error!(logger, "Error updating ledger metrics: {:?}", e);
+            }
         }
+
         std::thread::sleep(std::time::Duration::from_millis(1000));
     }
 }